edited operation func to accept decimal values

+++ main.js
@@ -76,28 +76,28 @@ buttons.forEach(button => {
 
         if (operator == "+") {
           console.log("return");
-          input.value = parseInt(valueA) + parseInt(valueB);
+          input.value = parseFloat(valueA) + parseFloat(valueB);
           valueA =  input.value;
           valueB = "";  //this will make sure the value of operator and valueB is empty string for the next calculation
           operator = ""; 
         } else if (operator == "-") {
-          input.value = parseInt(valueA) - parseInt(valueB);
+          input.value = parseFloat(valueA) - parseFloat(valueB);
           valueA =  input.value;
           valueB = "";
           operator = "";
         } else if (operator == "*") {
-          input.value = parseInt(valueA) * parseInt(valueB);
+          input.value = parseFloat(valueA) * parseFloat(valueB);
           valueA =  input.value;
           valueB = "";
           operator = "";
         } else if (operator == "÷") {
-          input.value = parseInt(valueA) / parseInt(valueB);
+          input.value = parseFloat(valueA) / parseFloat(valueB);
           valueA =  input.value;
           valueB = "";
           operator = "";
         }
         else if (operator == "%") {
-          input.value = (parseInt(valueA)/100) * parseInt(valueB);
+          input.value = (parseFloat(valueA)/100) * parseFloat(valueB);
           valueA =  input.value;
           valueB = "";
           operator = "";
@@ -106,9 +106,10 @@ buttons.forEach(button => {
 
 
       //separate the values entered into the inputbox by the operator in between them
-      // find a way to store the two values on which a given operator will act
+      // find a way to store the two values on which a given operator will act 
       // method for = , we have two values present otherwise do nothing
 
       //Problems: If we click multiple operators, the operator value is replaced and valueB is added to. 
-      // If the operator is typed before valueA, NaN is displayed (potential fix: not allow the operators to be entered before valueA)
+      // If the operator is typed before valueA, NaN is displayed (potential fix: not allow the operators to be entered before valueA or just NaN be displayed)
+      //decimal numbers are not working as decimals (potential fix: change parseInt, as it returns a whole number)
     })
